requirements.txt
transformers==4.46.3
accelerate
safetensors
qwen-vl-utils
Pillow
requests
tqdm
datasets
torch
matplotlib
numpy